History feature

+++ IssueTracker/IssueTrackerAPI/Controllers/ProjectController.cs
@@ -9,6 +9,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Models.Request;
 using Validation;
+using System.Security.Claims;
+using Models.Response;
 
 namespace IssueTrackerAPI.Controllers;
 
@@ -19,11 +21,16 @@ public class ProjectController : ControllerBase
     private readonly IProjectRepository _projectdb;
     private readonly IParticipantRepository _participantdb;
     private readonly Mapper _mapper;
-    public ProjectController(IProjectRepository projectdb, IParticipantRepository participantdb)
+    private readonly HistoryHandler _historyHandler;
+    private readonly IHistoryRepository _historyRepository;
+    public ProjectController(IProjectRepository projectdb, IParticipantRepository participantdb,
+        IHistoryRepository historyRepository)
     {
         _projectdb = projectdb;
         _participantdb = participantdb;
         _mapper = AutoMapperConfig.Config();
+        _historyHandler = new HistoryHandler(historyRepository);
+        _historyRepository = historyRepository;
     }
 
     [HttpPost("add-project")]
@@ -37,28 +44,53 @@ public class ProjectController : ControllerBase
             List<ValidationFailure> failures = result.Errors;
             return BadRequest(failures);
         }
+        var userclaim = User.Claims.FirstOrDefault(x => x.Type.Equals(ClaimTypes.Name));
+        if (userclaim == null) return Unauthorized();
         var project = _mapper.Map<Project>(entity);
         var res = await _projectdb.AddAsync(project);
         var userid = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
         var results = await ParticipantController.CreateOwner(res, Guid.Parse(userid!.Value), _participantdb);
-        if (results) return Ok();
+        if (results)
+        {
+            _historyHandler.CreatedProject(res, userclaim.Value, DateTime.UtcNow);
+            return Ok();
+        }
+
         return BadRequest();
     }
 
     [HttpGet("getAll-project")]
-    public async Task<IEnumerable<Project>> GetAll() =>
-        await _projectdb.GetAllAsync();
+    public async Task<IActionResult> GetAll()
+    {
+        var projects = await _projectdb.GetAllAsync();
+        if (projects.Any())
+        {
+            foreach (var project in projects)
+            {
+                project.History = await _historyRepository.GetByProjectIdAsync(project.Id);
+            }
+            IEnumerable<ProjectResponse> results = _mapper.Map<IEnumerable<ProjectResponse>>(projects);
+            return Ok(results);
+        }
+        return BadRequest();
+    }
 
     [HttpGet("getById-project")]
     public async Task<IActionResult> GetById(int id)
     {
         if (id == 0) return BadRequest("Invalid Id!");
-        return Ok(await _projectdb.GetByIdAsync(id));
+        var project = await _projectdb.GetByIdAsync(id);
+        if (project == null) return NotFound("Didn't find any project with this id");
+        project.History = await _historyRepository.GetByProjectIdAsync(project.Id);
+        var response = _mapper.Map<ProjectResponse>(project);
+        return Ok(response);
     }
     [HttpPut("update-project")]
     [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public async Task<IActionResult> Update(ProjectRequest entity)
     {
+        var userclaim = User.Claims.FirstOrDefault(x => x.Type.Equals(ClaimTypes.Name));
+        if (userclaim == null) return Unauthorized();
         var validator = new ProjectValidation();
         ValidationResult result = validator.Validate(entity);
         if (!result.IsValid)
@@ -77,8 +109,13 @@ public class ProjectController : ControllerBase
                     Guid.Parse(idclaim.Value), entity.Id);
             if (results == true)
             {
-                var project = _mapper.Map<Project>(entity);
-                await _projectdb.UpdateAsync(project);
+                var newProject = _mapper.Map<Project>(entity);
+                var oldProject = await _projectdb.GetByIdAsync(newProject.Id);
+                if (oldProject == null) return NotFound("There is no project with this id");
+                string field, newValue, oldValue;
+                await _projectdb.UpdateAsync(newProject);
+                _historyHandler.ProjectUpdatedValues(oldProject, newProject, out field, out oldValue, out newValue);
+                _historyHandler.UpdatedProject(newProject.Id, userclaim.Value, DateTime.UtcNow, field, oldValue, newValue);
                 return Ok();
             }
         }
@@ -86,17 +123,23 @@ public class ProjectController : ControllerBase
     }
 
     [HttpDelete("delete-project")]
-    public async Task<IActionResult> Update(int id)
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+    public async Task<IActionResult> Delete(int id)
     {
+        var userclaim = User.Claims.FirstOrDefault(x => x.Type.Equals(ClaimTypes.Name));
+        if (userclaim == null) return Unauthorized();
         if (id <= 0) return BadRequest("Invalid Id!");
         var idclaim = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
         if (idclaim != null)
         {
+            var project = await _projectdb.GetByIdAsync(id);
+            if (project == null) return NotFound("No project with this Id");
             var results = await CheckRole.IsOwner(_participantdb,
                     Guid.Parse(idclaim.Value), id);
             if (results == true)
             {
                 await _projectdb.DeleteAsync(id);
+                _historyHandler.DeletedProject(id, userclaim.Value, DateTime.UtcNow);
                 return Ok();
             }
         }
